requirements.txt
mysql-connector-python
PyMySQL
pandas
openpyxl
ortools
pytest
fastapi
uvicorn
python-multipart
black
yfinance
requests 
beautifulsoup4
